% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task-view.R
\name{extract_package_names}
\alias{extract_package_names}
\title{Extract package names from the xml for a .ctv formatted file}
\usage{
extract_package_names(xml)
}
\arguments{
\item{xml}{The xml content within a .ctv format file}
}
\description{
Extract package names from the xml for a .ctv formatted file
}
