% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task-view.R
\name{import_task_view_packages}
\alias{import_task_view_packages}
\title{Import a vector of package names from a "Task View" page}
\usage{
import_task_view_packages(url)
}
\arguments{
\item{url}{A filepath or URL containing a Task View page}
}
\value{
Character vector of package names
}
\description{
The file can either be a legacy `.ctv` file or a modern markdown-format task-view.
}
