% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task-view.R
\name{import_task_view_packages_ctv}
\alias{import_task_view_packages_ctv}
\title{Import a vector of package names from a (legacy) xml-format "Task View" page}
\usage{
import_task_view_packages_ctv(url)
}
\arguments{
\item{url}{A filepath or URL pointing to a .ctv file}
}
\value{
Character vector of package names
}
\description{
The old format of task-views used a ".ctv" file format, that was xml based.
}
