% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitsum.R
\name{run_gitsum_workflow}
\alias{run_gitsum_workflow}
\title{Run gitsum on an R package repository and store the results in a .tsv}
\usage{
run_gitsum_workflow(repo_path, output_path, package, r_dir_only = TRUE)
}
\arguments{
\item{repo_path}{Scalar character. Filepath to a git repository for the R package under
analysis.}

\item{output_path}{Scalar character. Filepath for storing the gitsum results as a `.tsv`.}

\item{package}{Scalar character. The name of the package that is being analysed.}

\item{r_dir_only}{Scalar logical. Should we restrict analysis to the `R/` source code
directory (default: TRUE)?}
}
\value{
tibble containing gitsum results for the repo (invisibly). This function is mainly
  called to set up a .tsv as a side-effect.
}
\description{
Run gitsum on an R package repository and store the results in a .tsv
}
